test: add SearchSG setup helper and integrate into SearchSGInputBox stories
