+++ packages/components/src/stories/helpers/index.ts
@@ -1,1 +1,2 @@
 export * from "./generateSiteConfig"
+export * from "./searchSg"
